backend/pyproject.toml
[project]
name = "pediatric-oncology-intelligence"
version = "0.1.0"
description = "Search intelligence pipeline for pediatric oncology trends analysis"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "openai>=1.12.0",
    "pytrends>=4.9.2",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "umap-learn>=0.5.5",
    "hdbscan>=0.8.33",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "azure-storage-blob>=12.19.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
